Move tests into the src folder and corresponding subfolders
#71
