+++ src/combinators/anyOf.test.ts
@@ -1,7 +1,7 @@
-import { anyOf } from "../../src/combinators/anyOf";
-import { many } from "../../src/combinators/many";
-import { string } from "../../src/parser/index";
-import { ParserState } from "../../src/ParserState";
+import { anyOf } from "./anyOf";
+import { many } from "./many";
+import { string } from "../parser/index";
+import { ParserState } from "../ParserState";
 
 describe('anyOf', () => {
   test('Parse a matchable input', () => {
