+++ src/combinators/sequenceOf.test.ts
