+++ src/parser/word.test.ts
